requirements.txt
Flask==0.12
gunicorn==19.6.0
google-cloud-pubsub==0.23.0
six==1.10.0

